refactor: UseCase 제거 및 아키텍처 구조 수정
