--- TIG/TIG/Domain/UseCase/TimeUseCase.swift
@@ -1,13 +1,0 @@
-//
-//  TimeUseCase.swift
-//  TIG
-//
-//  Created by Seo-Jooyoung on 7/26/24.
-//
-
-import SwiftUI
-
-@Observable
-class TimeUseCase {
-    
-}
