+++ TIG/TIG/Model/DTO/DailyContentSD.swift
