+++ TIG/TIG/Model/DTO/WeeklyRepeatSD.swift
