+++ TIG/TIG/Model/Entitiy/DailyContent.swift
