+++ TIG/TIG/Repository/Interface/AppSettingRepository.swift
