+++ TIG/TIG/Repository/Interface/DailyContentRepository.swift
